+++ .idea/misc.xml
@@ -17,6 +17,7 @@
         <entry key="app/src/main/res/drawable/baseline_logout_24.xml" value="0.3125" />
         <entry key="app/src/main/res/drawable/baseline_share_24.xml" value="0.3125" />
         <entry key="app/src/main/res/drawable/cancel_background.xml" value="0.2095" />
+        <entry key="app/src/main/res/drawable/cancel_button_background.xml" value="0.206" />
         <entry key="app/src/main/res/drawable/celadon_tag.xml" value="0.22291666666666668" />
         <entry key="app/src/main/res/drawable/celadon_toolbar.xml" value="0.1225" />
         <entry key="app/src/main/res/drawable/cicle.xml" value="0.3205" />
@@ -27,8 +28,8 @@
         <entry key="app/src/main/res/drawable/honey_dew.xml" value="0.22291666666666668" />
         <entry key="app/src/main/res/drawable/honey_drew_toolbar.xml" value="0.1225" />
         <entry key="app/src/main/res/drawable/ic_apricot_24.xml" value="0.353125" />
-        <entry key="app/src/main/res/drawable/ic_baseline_access_alarm_24.xml" value="0.179" />
-        <entry key="app/src/main/res/drawable/ic_baseline_access_time_24.xml" value="0.254" />
+        <entry key="app/src/main/res/drawable/ic_baseline_access_alarm_24.xml" value="0.15" />
+        <entry key="app/src/main/res/drawable/ic_baseline_access_time_24.xml" value="0.18" />
         <entry key="app/src/main/res/drawable/ic_baseline_add_24.xml" value="0.3125" />
         <entry key="app/src/main/res/drawable/ic_baseline_add_alert_24.xml" value="0.254" />
         <entry key="app/src/main/res/drawable/ic_baseline_add_location_24.xml" value="0.3125" />
@@ -50,7 +51,10 @@
         <entry key="app/src/main/res/drawable/ic_baseline_local_phone_24.xml" value="0.3545" />
         <entry key="app/src/main/res/drawable/ic_baseline_mode_edit_24.xml" value="0.2875" />
         <entry key="app/src/main/res/drawable/ic_baseline_more_vert_24.xml" value="0.307" />
+        <entry key="app/src/main/res/drawable/ic_baseline_palette_24.xml" value="0.206" />
         <entry key="app/src/main/res/drawable/ic_baseline_push_pin_24.xml" value="0.179" />
+        <entry key="app/src/main/res/drawable/ic_baseline_redo_24.xml" value="0.206" />
+        <entry key="app/src/main/res/drawable/ic_baseline_restore_24.xml" value="0.206" />
         <entry key="app/src/main/res/drawable/ic_baseline_search_24.xml" value="0.191" />
         <entry key="app/src/main/res/drawable/ic_baseline_settings_24.xml" value="0.307" />
         <entry key="app/src/main/res/drawable/ic_baseline_table_rows_24.xml" value="0.31" />
@@ -67,6 +71,7 @@
         <entry key="app/src/main/res/drawable/ic_outline_add_alert_24.xml" value="0.179" />
         <entry key="app/src/main/res/drawable/ic_outline_archive_24.xml" value="0.179" />
         <entry key="app/src/main/res/drawable/ic_outline_bookmark_add_24.xml" value="0.3545" />
+        <entry key="app/src/main/res/drawable/ic_outline_delete_24.xml" value="0.206" />
         <entry key="app/src/main/res/drawable/ic_outline_document_scanner_24.xml" value="0.195" />
         <entry key="app/src/main/res/drawable/ic_outline_push_pin_24.xml" value="0.179" />
         <entry key="app/src/main/res/drawable/icons_google.xml" value="0.3545" />
@@ -74,6 +79,7 @@
         <entry key="app/src/main/res/drawable/nav_acc_detail_background.xml" value="0.306" />
         <entry key="app/src/main/res/drawable/note_lock_24.xml" value="0.307" />
         <entry key="app/src/main/res/drawable/note_wrapper.xml" value="0.3205" />
+        <entry key="app/src/main/res/drawable/save_button_background.xml" value="0.206" />
         <entry key="app/src/main/res/drawable/save_shared_background.xml" value="0.2015" />
         <entry key="app/src/main/res/drawable/side_nav_bar.xml" value="0.1415" />
         <entry key="app/src/main/res/drawable/sign_in_bgnd.xml" value="0.299" />
@@ -125,7 +131,7 @@
         <entry key="app/src/main/res/layout/info_container.xml" value="0.14222222222222222" />
         <entry key="app/src/main/res/layout/label_fragment.xml" value="0.1" />
         <entry key="app/src/main/res/layout/label_item.xml" value="0.25" />
-        <entry key="app/src/main/res/layout/label_rv_item.xml" value="0.15" />
+        <entry key="app/src/main/res/layout/label_rv_item.xml" value="0.5" />
         <entry key="app/src/main/res/layout/label_title_dialog.xml" value="0.1" />
         <entry key="app/src/main/res/layout/main_action_bar.xml" value="0.22" />
         <entry key="app/src/main/res/layout/main_bar_back.xml" value="0.24322916666666666" />
@@ -136,7 +142,7 @@
         <entry key="app/src/main/res/layout/note_back_bottom_sheet.xml" value="0.47232472324723246" />
         <entry key="app/src/main/res/layout/note_label_bottom_sheet.xml" value="0.25" />
         <entry key="app/src/main/res/layout/note_rv_item.xml" value="0.17803030303030304" />
-        <entry key="app/src/main/res/layout/save_shared_note_fragment.xml" value="0.15" />
+        <entry key="app/src/main/res/layout/save_shared_note_fragment.xml" value="0.2966101694915254" />
         <entry key="app/src/main/res/layout/settings_fragment.xml" value="0.33" />
         <entry key="app/src/main/res/layout/spinner_tag_row.xml" value="0.3328125" />
         <entry key="app/src/main/res/layout/tag_adapter_item.xml" value="0.75" />
@@ -159,7 +165,7 @@
       </map>
     </option>
   </component>
-  <component name="ProjectRootManager" version="2" languageLevel="JDK_X" default="true" project-jdk-name="Android Studio default JDK" project-jdk-type="JavaSDK">
+  <component name="ProjectRootManager" version="2" languageLevel="JDK_16" project-jdk-name="Android Studio default JDK" project-jdk-type="JavaSDK">
     <output url="file://$PROJECT_DIR$/build/classes" />
   </component>
   <component name="ProjectType">
