+++ app/build.gradle
@@ -21,7 +21,7 @@ android {
         applicationId "com.neuralbit.letsnote"
         minSdk 25
         targetSdk 33
-        versionCode 12
+        versionCode 13
         versionName "1.0"
 
         testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
