+++ app/src/main/AndroidManifest.xml
@@ -21,7 +21,7 @@
         android:label="@string/app_name"
         android:roundIcon="@mipmap/ic_launcher1_round"
         android:supportsRtl="true"
-        android:theme="@style/Theme.App.Starting">
+        android:theme="@style/Theme.App.Activity">
         <activity
             android:name=".ui.signIn.PrivacyPolicy"
             android:exported="false"
