+++ app/src/main/java/com/neuralbit/letsnote/ui/addEditNote/AddEditNoteActivity.kt
@@ -678,7 +678,7 @@ class AddEditNoteActivity : AppCompatActivity() ,
             if (noteContentSplit[lineIndex].endsWith(":")) {
                 if (noteContent.endsWith("\n")) {
                     if (!backPressed) {
-                        noteDescriptionEdit.append("-> ")
+                        noteDescriptionEdit.append("- ")
 
                     }
                 }
@@ -1646,14 +1646,14 @@ class AddEditNoteActivity : AppCompatActivity() ,
         this.calendar[Calendar.HOUR]= calendar[Calendar.HOUR]
         this.calendar[Calendar.MINUTE]= calendar[Calendar.MINUTE]
         this.calendar[Calendar.SECOND]= calendar[Calendar.SECOND]
-        timeTitleTV.text= resources.getString(R.string.date_time_set,"Time set ", DateFormat.getTimeFormat(applicationContext).format(calendar.time))
+        timeTitleTV.text= resources.getString(R.string.date_time_set,"Time ", DateFormat.getTimeFormat(applicationContext).format(calendar.time))
     }
 
     override fun getDateInfo(calendar : Calendar) {
         this.calendar[Calendar.DAY_OF_MONTH] = calendar[Calendar.DAY_OF_MONTH]
         this.calendar[Calendar.MONTH] = calendar[Calendar.MONTH]
         this.calendar[Calendar.YEAR] = calendar[Calendar.YEAR]
-        dateTitleTV.text= resources.getString(R.string.date_time_set,"Date set ", DateFormat.getDateFormat(applicationContext).format(calendar.time))
+        dateTitleTV.text= resources.getString(R.string.date_time_set,"Date ", DateFormat.getDateFormat(applicationContext).format(calendar.time))
     }
 
     override fun getTag(tag: String) {
