+++ app/src/main/res/drawable/ic_baseline_access_alarm_note_rv_24.xml
@@ -1,4 +1,5 @@
 <vector xmlns:android="http://schemas.android.com/apk/res/android"
+
     android:width="24dp"
     android:height="24dp"
     android:viewportWidth="24"
