+++ app/src/main/res/drawable/save_button_background.xml
@@ -1,0 +1,11 @@
+<?xml version="1.0" encoding="utf-8"?>
+<shape
+    xmlns:android="http://schemas.android.com/apk/res/android"
+    android:shape="rectangle"
+    >
+
+    <corners
+        android:radius="50dp"/>
+    <solid
+        android:color="@color/blush"/>
+</shape>
