+++ app/src/main/res/layout/activity_add_edit_note.xml
@@ -53,6 +53,7 @@
                     android:layout_width="match_parent"
                     android:layout_height="wrap_content"
                     android:autoLink="web"
+                    android:layout_marginTop="8dp"
                     android:background="@android:color/transparent"
                     android:completionThreshold="1"
                     android:gravity="top"
@@ -158,7 +159,7 @@
             android:visibility="gone"
             android:layout_margin="@dimen/margin"
             android:background="?attr/selectableItemBackgroundBorderless"
-            android:src="@drawable/ic_baseline_access_alarm_24"
+            android:src="@drawable/ic_baseline_access_alarm_add_edit_note_24"
             app:layout_constraintStart_toStartOf="parent"
             app:layout_constraintBottom_toTopOf="@id/infoContainer"
             android:contentDescription="@string/reminder_icon" />
