+++ app/src/main/res/layout/save_shared_note_fragment.xml
@@ -6,7 +6,7 @@
     android:layout_gravity="center"
     android:padding="16dp"
     android:background="@drawable/save_shared_background"
-    android:layout_height="500dp">
+    android:layout_height="wrap_content">
     <de.hdodenhof.circleimageview.CircleImageView
         android:layout_width="30dp"
         android:layout_height="30dp"
@@ -45,28 +45,31 @@
         android:lineSpacingMultiplier="1.1"
         android:maxLines="8"
         android:textSize="16sp"
-        app:layout_constraintBottom_toTopOf="@+id/saveNoteButton"
+        android:minHeight="60dp"
+        android:layout_marginTop="8dp"
+        app:layout_goneMarginTop="8dp"
         app:layout_constraintTop_toBottomOf="@id/noteEditTitle"
-        tools:layout_editor_absoluteX="16dp"
         android:inputType="text"
         android:importantForAutofill="no" />
+
     <Button
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
         android:id="@+id/cancelButton"
-        android:text="@string/Cancel"
-        android:textColor="@color/timberwolf"
-        android:backgroundTint="@color/gunmetal"
+        android:layout_width="wrap_content"
+        android:layout_height="35dp"
         android:layout_marginEnd="8dp"
+        android:background="@drawable/cancel_button_background"
+        android:text="@string/Cancel"
+        android:layout_marginTop="16dp"
+        android:textColor="@color/timberwolf"
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toStartOf="@id/saveNoteButton"
-        />
+        app:layout_constraintTop_toBottomOf="@+id/noteEditDesc" />
     <Button
         android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
+        android:layout_height="35dp"
         android:id="@+id/saveNoteButton"
         android:text="@string/Save"
-        android:backgroundTint="@color/blush"
+        android:background="@drawable/save_button_background"
         android:textColor="@color/timberwolf"
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
